review: address codex + claude-bot PR feedback
Codex P1 — SMT soundness for dropped flag-writers:
mcmc.rs:199 and synthesis.rs:238 call check_equivalence_with_config with a
register-only LiveOutMask, while the SMT models ADDS/SUBS/ANDS/NEGS/BICS
identically to their non-flag siblings. A rewrite that drops a flag-writer
would therefore be certified equivalent in registers even though a
downstream CSEL / B.cond would observe different NZCV. Adds a
drops_target_flag_writer() pre-SMT guard inside check_equivalence,
check_equivalence_with_config, and check_equivalence_with_config_metrics
that returns NotEquivalent whenever the target writes flags but the
candidate does not. Regression test
test_adds_to_add_rewrite_rejected_even_with_register_live_out covers both
entry points.

Codex P2 — exhaustive symbolic generator coverage:
generate_all_instructions in src/search/candidate.rs only emitted the
original MOV / ALU / shift set; the new Tier 1 variants (MVN, NEG, NEGS,
MOVN, BIC, BICS, ORN, EON, ADDS, SUBS, ANDS, CSET, CSETM, ROR) are now
enumerated too, with a small representative MOVN imm / shift table to
avoid blowing up the candidate count.

Clippy (github-advanced-security):
- src/semantics/concrete.rs:192 — drop the redundant `as u64` cast.
- src/semantics/smt.rs:243 — same redundant cast.
- src/semantics/smt.rs:249/255/261 — drop the unnecessary `&` borrows
  on `rhs.bvnot()`; bvand/bvor/bvxor accept the owned BV directly.

claude-bot review summary (item-by-item):
1. Consolidated the three duplicate random_normal_cond helpers in
   isa/aarch64.rs, search/candidate.rs, and stochastic/mutation.rs into a
   single Condition::random_normal() backed by a NORMAL_CONDITIONS
   constant in ir/types.rs. generate_all_instructions also reuses the
   constant.
2. Added a comment on the ROR SMT formula explaining the SMTLIB2
   bvshl(_, 64) = 0 invariant — adding a guard for n == 0 would
   mis-handle symbolic n.
3. Added a comment on the BIC | Bics SMT arm pointing at
   drops_target_flag_writer as the soundness barrier for BICS's flag
   side-effect.
4. (DEFERRED to a follow-up issue) The "operand mutation can produce
   out-of-range immediates for ADD / SUB / ADDS / SUBS" point is a
   pre-existing behaviour of mutate_operand; the search is robust to it
   (invalid candidates are filtered). Tracking issue to be filed.
5. Added a comment on the MOVN arm of aarch64.rs mutate explaining that
   the "change source operand" strategy only mutates `imm` by design,
   while the per-field mutator in stochastic/mutation.rs covers `shift`.
6. Subsumed by the clippy redundant-cast fix above.
7. Renamed test_movn_zero_equivalent_to_csetm_al to
   test_movn_zero_equivalent_to_eon_self — the original name referred to
   a CSETM/AL form that the test never builds.
8. Deduplicated the identical pick_rn / pick_reg closures in
   generate_random_instruction.
9. Removed the now-unnecessary #[allow(dead_code)] on Condition::invert
   — it's called by the CSET / CSETM encoders.

Tests: 315 → 317. CI check (fmt, build, unit + integration, test binaries,
full suite) passes.

+++ src/semantics/concrete.rs
@@ -190,7 +190,7 @@ pub fn apply_instruction_concrete(
         }
         // MOVN: rd = !((imm as u64) << shift)
         Instruction::MovN { rd, imm, shift } => {
-            let value = !(((*imm as u64) << (*shift as u32)) as u64);
+            let value = !((*imm as u64) << (*shift as u32));
             state.set_register(*rd, ConcreteValue::new(value));
         }
         // BIC: rd = rn & !rm
